requirements.txt
asgiref==3.5.0
attrs==21.4.0
Beaker==1.11.0
CacheControl==0.12.10     
cachy==0.3.0
certifi==2021.10.8        
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
ClickUz==0.3
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==36.0.2      
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0    
dj-email-url==1.0.5       
Django==4.0.3
django-cache-url==3.4.0
django-ckeditor==6.3.0
django-cors-headers==3.10.1
django-countries==7.3.2
django-environ==0.8.1
django-filter==21.1
django-forestadmin==1.2.0
django-ipware==4.0.2
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
drf-spectacular-sidecar==2022.4.1
drf-yasg==1.20.0
ecdsa==0.17.0
environs==9.5.0
filelock==3.6.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
keyring==23.5.0
lockfile==0.12.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.14.1
marshmallow-jsonapi==0.24.0
msgpack==1.0.3
numpy==1.22.3
oic==1.3.0
packaging==20.9
pandas==1.4.2
pastel==0.2.1
pexpect==4.8.0
Pillow==9.1.0
pkginfo==1.8.2
platformdirs==2.5.1
poetry==1.1.13
poetry-core==1.0.8
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.14.1
pyjwkest==1.4.2
PyJWT==2.3.0
pylev==1.4.0
pyparsing==3.0.7
pymemcache==3.5.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
tomlkit==0.10.1
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
webencodings==0.5.1
whitenoise==6.0.0
zipp==3.8.0